+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 v2.0.0
 
 * Remove usage of `sendAction` on the `tour-start-button` component - use closure actions instead
+* Ensure all callouts are removed when the tour-manager service is destroyed
 
 v1.1.4